+++ __init__.py
@@ -1,20 +1,4 @@
 """fortigate-api"""
 
-from requests import Session
-
 from fortigate_api.fortigate import Fortigate
-from fortigate_api.address import Address
-from fortigate_api.address_group import AddressGroup
-from fortigate_api.antivirus import Antivirus
-from fortigate_api.application import Application
-from fortigate_api.interface import Interface
-from fortigate_api.internet_service import InternetService
-from fortigate_api.ip_pool import IpPool
-from fortigate_api.policy import Policy
-from fortigate_api.schedule import Schedule
-from fortigate_api.service import Service
-from fortigate_api.service_category import ServiceCategory
-from fortigate_api.service_group import ServiceGroup
-from fortigate_api.snmp_community import SnmpCommunity
-from fortigate_api.virtual_ip import VirtualIP
-from fortigate_api.zone import Zone
+from fortigate_api.fortigate_api import FortigateAPI